pyproject.toml
[project]
license = "GPL-3.0-or-later"
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Text Processing :: General",
]
name = "recapit"
version = "0.17.0"
requires-python = ">=3.10"
dependencies = [
  "rich>=13.7",
  "natsort>=8.4",
  "pdf2image>=1.17",
  "PyPDF2>=3.0",
  "pillow>=10.0",
  "pillow-avif-plugin>=1.4",
  "typer>=0.12",
  "google-genai>=0.2",
  "PyYAML>=6.0",
  "yt-dlp>=2024.10.07",
]

[project.scripts]
recapit = "recapit.cli:main"

[tool.setuptools]
packages = ["recapit"]
